DCT_Steganography/requirements.txt
numpy>=1.24.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
scikit-image>=0.21.0
matplotlib>=3.7.0
cryptography>=41.0.0
reedsolo>=1.7.0
Pillow>=10.0.0
scipy>=1.11.0
pandas>=2.0.0

